// src/core/enums/SecsItemType.ts
export enum SecsItemType {
  L = 0x00, // List
  B = 0x20, // Binary
  BOOLEAN = 0x24, // Boolean
  A = 0x40, // ASCII
  I8 = 0x60, // Integer 8
  I1 = 0x64, // Integer 1
  I2 = 0x68, // Integer 2
  I4 = 0x70, // Integer 4
  F8 = 0x80, // Floating Point 8
  F4 = 0x90, // Floating Point 4
  U8 = 0xa0, // Unsigned Integer 8
  U1 = 0xa4, // Unsigned Integer 1
  U2 = 0xa8, // Unsigned Integer 2
  U4 = 0xb0, // Unsigned Integer 4
}